requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.5

# Language Models
llama-cpp-python>=0.2.0
openai>=1.0.0

# Vector Store
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Text Processing
tiktoken>=0.5.0
python-dotenv>=1.0.0

# Utils
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.2.0
pathlib>=1.0.1

# Optional but recommended
jupyter>=1.0.0
pandas>=2.0.0 